Innovation Team reactable Theme

Contains the parameters and colour for a standarized table theme.

Usage

react_format

Format

An object of class reactableTheme of length 13.

Value

an object with lists to pass into the reactable function or options.

Author

Alejandro Abraham a.abraham@innovationteam.eu

Examples

library(innteamUtils)
library(reactable)
reactable(iris, theme = react_format,
         highlight = TRUE, striped = TRUE, filterable = TRUE, resizable = TRUE, wrap = FALSE)